created controller for order
