+++ api/middleware/user_Auth.js
@@ -4,7 +4,7 @@ const jwt = require('jsonwebtoken');
 module.exports = (req, res, next ) =>{
     try{
         const token = req.headers.authorization.split(" ")[1];
-       // console.log(token);
+        console.log(token);
         const decoder = jwt.verify(token, process.env.JWT_KEY)
         req.userData = decoder;
         next();
