+++ api/routes/order.js
@@ -2,116 +2,19 @@ const express = require('express');
 const router = express.Router();
 const mongoose = require('mongoose');
 const Order = require('../models/orderSchema')
-const Menu = require('../models/foodMenu')
+const Menu = require('../models/foodMenu');
 const checkAuth = require('../middleware/user_Auth');
+const orderController = require('../controllers/order');
 
 
-router.get('/', checkAuth, (req, res, next)=>{
-    Order.find()
-    .select("foodId quantity _id")
-    .exec()
-    .then(docs=>{
-        console.log(docs);
-     const response = {
-     count :docs.length, 
-      orders: docs.map(doc=>{
-          return{
-         foodId: doc.foodId,
-         _id  : doc._id,
-         quantity: doc.quantity,
-         request :{type: 'GET', url: 'http://localhost:5000/order/' + doc._id} 
-          }
-      })
-    }
-  
-    res.status(200).json({response})
-
-})
-    .catch(err => {console.log(err)
-        res.status(201).json({error: err})
-    });
-});
+router.get('/', checkAuth, orderController.orders_get_all)
 
 
-router.post('/',checkAuth , (req, res, next)=>{
-    Menu.findById(req.body.foodId)
-    .then( food =>{
-        if(!food){
-            return res.status(404).json({message: "Food not found"})
-        }
-        const order  = new Order({
-            _id : mongoose.Types.ObjectId(),
-            quantity: req.body.quantity,
-            foodId: req.body.foodId
-        });
-        return order
-        .save().then(result=>{
-            console.log(result);
-            res.status(201).json({
-                message: 'order was created',
-                createdProperty: {
-                    _id : result._id,
-                    foodId: result.foodId,
-                    quantity: result.quantity,
-                    request :{
-                        type :"GET",
-                        url:"http://localhost:5000/order/" + result._id
-                    }
-               }
-           })
-        })
-    })
-    
-    .catch(err=>{
-        console.log(err);
-        res.status(500).json({error: err});
-        });
-     });
+router.post('/',checkAuth , orderController.order_create_orders);
     
 
+router.get('/:orderId', checkAuth, orderController.order_get_order);
 
-
-router.get('/:orderId', checkAuth, (req, res, next)=>{
-    const id = req.params.orderId;
-Order.findById(id)
-.exec()
-.then(order =>{
-    if(!order){
-        return res.status(404).json({message: "order not found"});
-    }
-    console.log(order)
-res.status(200).json({
-    Order: order,
-    request:{
-        type : 'GET',
-        description: 'GET_ALL_ORDER',
-        url : "http://localhost:5000/order" 
-    }
-})
-})
-.catch(err =>{console.log(err)
-res.status(500).json({error:err})
-})
-});   
-router.delete('/:orderId', checkAuth, (req, res, next)=>{
-    const id = req.params.orderId;
-    Order.remove({_id : id})
-    .exec()
-    .then(result => {
-     console.log(result)
-     res.status(200).json({
-     message : "order deleted",
-     request: {
-      type:"POST",
-      url : "http://localhost:5000/order" ,
-      body: {foodId: "ID", quantity: "Number"}
-     }
-     })
-    })
-    .catch(err =>{
-    console.log(err)
-    res.status(500).json({error: err})
-    });
-});
-
+router.delete('/:orderId', checkAuth, orderController.order_delete_order)
+   
 module.exports = router;
